% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\name{WebTRIS_clean_sites}
\alias{WebTRIS_clean_sites}
\title{Get Cleans Site Names
The Name field contains useful data such as the direction of flow, this function extracts this data.}
\usage{
WebTRIS_clean_sites(x)
}
\arguments{
\item{x}{Data.Frame from WebTRIS_sites()}
}
\description{
Get Cleans Site Names
The Name field contains useful data such as the direction of flow, this function extracts this data.
}
\details{

}
