% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{WebTRIS_reports}
\alias{WebTRIS_reports}
\title{Get reports}
\usage{
WebTRIS_reports(sites = NA, period = "daily", start_date = Sys.Date()
  - 60, end_date = Sys.Date())
}
\arguments{
\item{sites}{Number of the WebTRIS site e.g. 8188}

\item{period}{Time period of reports can be "monthly" or "daily". Default "daily"}

\item{start_date}{Start date as object of class Date. Default Today - 60 days}

\item{end_date}{End date as object of class Date. Default Today}
}
\description{
Get reports
}
\details{
This function returns a report for the given WebTRIS site based
}
