% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\name{WebTRIS_sites}
\alias{WebTRIS_sites}
\title{Get sites}
\usage{
WebTRIS_sites(sites = NA, clean = TRUE)
}
\arguments{
\item{sites}{Number vector of site ids or NA for all sites}
}
\description{
Get sites
}
\details{
This function returns sites data for the given WebTRIS site based
}
